Fix floordiv by avoiding rounding up due to low precision floats

+++ src/runtime/lpython_builtin.py
@@ -440,8 +440,8 @@ def _lpython_floordiv(a: f64, b: f64) -> f64:
 
 @overload
 def _lpython_floordiv(a: f32, b: f32) -> f32:
-    r: f32
-    r = a/b
+    r: f64
+    r = float(a)/float(b)
     result: i32
     resultf32: f32
     result = int(r)
@@ -454,7 +454,7 @@ def _lpython_floordiv(a: f32, b: f32) -> f32:
 @overload
 def _lpython_floordiv(a: i32, b: i32) -> i32:
     r: f64 # f32 rounds things up and gives incorrect results
-    r = a/b
+    r = float(a)/float(b)
     result: i32
     result = int(r)
     if r >= 0.0 or result == r:
